superhack-backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
passlib[bcrypt]
pyjwt
pytest
httpx
